requirements.txt
tensorflow>=2.0
opencv-python-headless
pandas
numpy
Pillow
matplotlib
scikit-learn
seaborn
ipykernel
jupyter
